refactor: move partition key validation to segment_store/utils.py
Mirrors the vector store's layout (validate_identifier in
vector_store/utils.py); the interface docstring states the key rule
plainly instead of referencing a code path.
