+++ packages/server/src/memmachine_server/episodic_memory/event_memory/segment_store/segment_store.py
@@ -151,10 +151,7 @@ class SegmentStore(ABC):
 
         Args:
             partition_key (str):
-                The key of the partition. Keys are embedded in native
-                storage identifiers, so they must be lowercase
-                alphanumeric/underscore and at most 32 characters
-                (see data_types.validate_partition_key).
+                The key of the partition (a-z0-9_, max 32 characters).
             config (SegmentStorePartitionConfig):
                 Configuration for the partition.
