+++ packages/server/src/memmachine_server/episodic_memory/event_memory/segment_store/utils.py
@@ -1,0 +1,21 @@
+"""Shared utilities for segment store implementations."""
+
+import re
+
+_PARTITION_KEY_RE = re.compile(r"^[a-z0-9_]+$")
+_PARTITION_KEY_MAX_LENGTH = 32
+
+
+def validate_partition_key(partition_key: str) -> None:
+    """Raise ValueError unless the key is valid (a-z0-9_, max 32 characters)."""
+    if not _PARTITION_KEY_RE.match(partition_key):
+        raise ValueError(
+            f"Partition key {partition_key!r} contains invalid characters. "
+            "Only lowercase alphanumeric and underscores are allowed."
+        )
+    if len(partition_key) > _PARTITION_KEY_MAX_LENGTH:
+        raise ValueError(
+            f"Partition key {partition_key!r} is too long "
+            f"({len(partition_key)} characters). "
+            f"Maximum is {_PARTITION_KEY_MAX_LENGTH}."
+        )
